update. ルーティングをresourceにまとめました。
